+++ src/JavaScriptRuntime/Node/Contracts/IFsModule.Generated.cs
@@ -10,11 +10,7 @@ namespace Jroc.Runtime.Node.Contracts;
 /// <summary>
 /// Defines the public top-level <c>node:fs</c> module contract from Node.js 24.18.1.
 /// </summary>
-/// <remarks>
-/// Nested option, result, and handle contracts intentionally remain dynamic in this proof of concept.
-/// They will be strongly typed by the work tracked in GitHub issue #1660.
-/// </remarks>
-[global::System.CodeDom.Compiler.GeneratedCode("generateNodeModuleInterface.js", "sha256:e0d88d11483936ce8b947360e0bac6d198acd445d3e7638f60398a795de2bba7")]
+[global::System.CodeDom.Compiler.GeneratedCode("generateNodeModuleInterface.js", "sha256:13ea61f00a6a5629ba94c535a3f32f6ddcc06ed57c5bde2aa5bdbf56d257d242")]
 [NodeModuleInterface("fs")]
 public interface IFsModule
 {
@@ -296,18 +292,21 @@ public interface IFsModule
     /// <summary>
     /// Node.js signature: <c>fs.open(path[, flags[, mode]], callback)</c>.
     /// </summary>
+    [global::Jroc.Runtime.Node.Contracts.NodeModuleResultContract(global::Jroc.Runtime.Node.Contracts.NodeModuleResultKind.Callback, typeof(global::Jroc.Runtime.Node.Contracts.IFsFileHandle), "callback")]
     [NodeModuleMember("open")]
     void open(object? path, global::System.Delegate callback);
 
     /// <summary>
     /// Node.js signature: <c>fs.open(path[, flags[, mode]], callback)</c>.
     /// </summary>
+    [global::Jroc.Runtime.Node.Contracts.NodeModuleResultContract(global::Jroc.Runtime.Node.Contracts.NodeModuleResultKind.Callback, typeof(global::Jroc.Runtime.Node.Contracts.IFsFileHandle), "callback")]
     [NodeModuleMember("open")]
     void open(object? path, object? flags, global::System.Delegate callback);
 
     /// <summary>
     /// Node.js signature: <c>fs.open(path[, flags[, mode]], callback)</c>.
     /// </summary>
+    [global::Jroc.Runtime.Node.Contracts.NodeModuleResultContract(global::Jroc.Runtime.Node.Contracts.NodeModuleResultKind.Callback, typeof(global::Jroc.Runtime.Node.Contracts.IFsFileHandle), "callback")]
     [NodeModuleMember("open")]
     void open(object? path, object? flags, object? mode, global::System.Delegate callback);
 
@@ -381,7 +380,7 @@ public interface IFsModule
     /// Node.js signature: <c>fs.readFile(path[, options], callback)</c>.
     /// </summary>
     [NodeModuleMember("readFile")]
-    void readFile(object? path, object? options, global::System.Delegate callback);
+    void readFile(object? path, [global::Jroc.Runtime.Node.Contracts.NodeModuleParameterContract(typeof(global::Jroc.Runtime.Node.Contracts.IFsReadFileOptions))] object? options, global::System.Delegate callback);
 
     /// <summary>
     /// Node.js signature: <c>fs.readlink(path[, options], callback)</c>.
@@ -621,7 +620,7 @@ public interface IFsModule
     /// Node.js signature: <c>fs.writeFile(file, data[, options], callback)</c>.
     /// </summary>
     [NodeModuleMember("writeFile")]
-    void writeFile(object? file, object? data, object? options, global::System.Delegate callback);
+    void writeFile(object? file, object? data, [global::Jroc.Runtime.Node.Contracts.NodeModuleParameterContract(typeof(global::Jroc.Runtime.Node.Contracts.IFsWriteFileOptions))] object? options, global::System.Delegate callback);
 
     /// <summary>
     /// Node.js signature: <c>fs.writev(fd, buffers[, position], callback)</c>.