+++ tests/Jroc.Tests/Node/JavaScript/PerfHooks_DestructuredPerformanceOverride.js
@@ -1,0 +1,11 @@
+"use strict";
+
+const perfHooks = require("perf_hooks");
+perfHooks.performance = Object.create({
+    now() {
+        return 42;
+    }
+});
+
+const { performance } = require("perf_hooks");
+console.log(performance.now());